refactor: rename observationsComponents to observationsHelpers
