+++ src/renderer/components/ObservationCard.tsx
@@ -2,7 +2,7 @@ import { Card, Elevation, Tooltip } from '@blueprintjs/core';
 import { useTranslation } from 'react-i18next';
 
 import styles from './ObservationCard.module.scss';
-import { getPredictions, PhotoDetails, PredictionsTable } from './observationsComponents';
+import { getPredictions, PhotoDetails, PredictionsTable } from './observationsHelpers';
 
 export default function ObservationCard(props: ObservationCardProps) {
   const {
